Fixed #26009 -- Fixed contenttypes_tests isolation.
